2018_SCTF/Final/attack/Millionaire/deploy/requirements.txt
aiohttp==2.3.10
asn1crypto==0.24.0
async-lru==0.1.0
async-timeout==3.0.0
attrdict==2.0.0
backcall==0.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coincurve==7.1.0
cryptography==2.2.2
cytoolz==0.9.0.1
decorator==4.3.0
eth-abi==1.1.1
eth-account==0.2.2
eth-bloom==1.0.0
eth-hash==0.1.3
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-tester==0.1.0b24
eth-utils==1.0.3
hexbytes==0.1.0
idna==2.6
idna-ssl==1.0.1
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
lru-dict==1.1.6
multidict==4.3.1
parsimonious==0.8.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
plyvel==1.0.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py-ecc==1.4.2
py-evm==0.2.0a16
py-geth==2.0.1
py-solc==3.0.0
pycparser==2.18
pycryptodome==3.6.1
pyethash==0.1.27
Pygments==2.2.0
pysha3==1.0.2
requests==2.18.4
rlp==1.0.1
semantic-version==2.6.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.9.0
traitlets==4.3.2
trie==1.3.6
urllib3==1.22
wcwidth==0.1.7
web3==4.2.1
websockets==5.0
yarl==1.2.4
